Added support for using BERT for mask filling.
This task does not work perfectly with the intentions and assumptions in the code from the text generation capabilities,
but I added it in a way that I think is reasonably maintainable and extensible. This works by sending prompts which are just
complete sentences that precede the prompt for which you want inference, to give the model more context.